fix(unity): instantiate local uint -> ulong (missed by local compile gates)
LoomStageDriver.Instantiate kept a uint local for the instantiate return
(NodeId is u64 since #26; RootSentinel is ulong.MaxValue — CS0266 +
CS0652 in Unity only). The headless projects never compile top-level
Runtime files (UnityEngine refs; they link Public/Projection/Bindings
only), so the u64 sweep's green gates could not see it. AGENTS compile-
gate note sharpened with the exact boundary and the hand-grep rule for
ABI-width sweeps.

+++ unity/package/Runtime/LoomStageDriver.cs
@@ -387,7 +387,7 @@ namespace LoomGUI
             StageHandle* h = (StageHandle*)_host.StagePtr.ToPointer();
             byte[] pb = Encoding.UTF8.GetBytes(pkgName ?? "");
             byte[] cb = Encoding.UTF8.GetBytes(compPath ?? "");
-            uint instId;
+            ulong instId;
             fixed (byte* pp = pb)
             fixed (byte* cp = cb)
                 instId = Native.loomgui_stage_instantiate(h, pp, (nuint)pb.Length, cp, (nuint)cb.Length);
